chore: move StockDTO to DTO folder
