+++ stock-server/src/main/java/donTouch/stock_server/calendar/dto/CalendarForm.java
@@ -1,0 +1,16 @@
+package donTouch.stock_server.calendar.dto;
+
+import jakarta.validation.constraints.NotBlank;
+
+import java.util.Date;
+
+public class CalendarForm {
+    @NotBlank(message = "input token")
+    String token;
+
+    @NotBlank(message = "input startDate")
+    Date startDate;
+
+    @NotBlank(message = "input endDate")
+    Date endDate;
+}
