+++ stock-server/src/main/java/donTouch/stock_server/stock/service/StockService.java
@@ -1,7 +1,7 @@
 package donTouch.stock_server.stock.service;
 
-import donTouch.stock_server.stock.domain.StockDTO;
 import donTouch.stock_server.stock.dto.FindStockForm;
+import donTouch.stock_server.stock.dto.StockDTO;
 
 import java.util.List;
